put TimeAgo func in another util file
